+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Extensions
 - `CGSize.withReversedDimensions()` and `CGSize.withReversedDimensions(if:)` methods are added
 - `Array.reversed(if:)` method is added that reverses elements on a condition
 - `Numeric.withOppositeSign()` and `Numeric.withOppositeSign(if:)` methods are added
+- `Range.reversedArray(if:)` and `ClosedRange.reversedArray(if:)` methods are added returns `Array` with reversed elements on a condition
 - Half-open range clamped methods have been removed due to an issue and an impossible implementation
 
 ### [4.4.1(70)](https://github.com/VakhoKontridze/VCore/releases/tag/4.4.1) — *2023 02 10*